refactor(react): group files by layer; colocate tests in __tests__
Move turnkey components into components/ and dnd glue into dnd/ (dragFlow + dragData, the renamed ex-context.ts). Unit tests now sit in __tests__/ next to their subject; cross-module integration tests live in src/__tests__/. Import-path changes only; no behavior change.
